chore: refactor folder hierarchy
