+++ config/paths.js
@@ -55,7 +55,8 @@ module.exports = {
   appBuild: resolveApp('build'),
   appPublic: resolveApp('public'),
   appHtml: resolveApp('public/index.html'),
-  appIndexJs: resolveModule(resolveApp, 'src/index'),
+  popupIndexJs: resolveModule(resolveApp, 'src/popup/index'),
+  contentScriptIndexJs: resolveModule(resolveApp, 'src/content-script/index'),
   appPackageJson: resolveApp('package.json'),
   appSrc: resolveApp('src'),
   appTsConfig: resolveApp('tsconfig.json'),
