+++ public/manifest.json
@@ -13,8 +13,8 @@
   "content_scripts": [
     {
       "matches": ["*://*.youtube.com/watch*"],
-      "css": ["/static/css/content.css"],
-      "js": ["static/js/content.js"]
+      "css": ["/static/css/contentScript.css"],
+      "js": ["static/js/contentScript.js"]
     }
   ],
   "icons": {
